Create README.md
Simon Game - Test Your Memory Skills!

Challenge your memory with the Simon Game! Test your ability to recall and replicate sequences of flashing colours in this classic memory game brought to life in a web-based format. With each level, the game becomes more challenging as additional colours are added to the sequence.